requirements.txt
discord.py

#(built in)
os 
